ミスってたところを直した

+++ linux_6.1_mod/kernel/signal.c
@@ -3204,8 +3204,9 @@ error_sigprocmask:
 success:
 	retval = 0;
 	// sigset_tはunsigned long(%lu)で処理する
-	sigset_msg_len = snprintf(sigset_msg_buf, 200, "%lu%c%lu", new_set,
-				  SCLDA_DELIMITER, old_set);
+	sigset_msg_len = snprintf(sigset_msg_buf, 200, "%lu%c%lu",
+				  (unsigned long)new_set, SCLDA_DELIMITER,
+				  (unsigned long)old_set);
 	goto sclda;
 sclda:
 	if (!is_sclda_allsend_fin())
